Add gameover message and sorting for arrays

+++ BattleArena/rand.cpp
@@ -13,7 +13,7 @@ int setRandomAttackValue(int minValue, int maxValue)
 }
 
 
-int setRandomIndex(int maxPeople)
+int setRandomIndex(size_t maxPeople)
 {
 	
 	return rand() % maxPeople;
